+++ packages/react-generators/src/generators/core/react-tailwind/index.ts
@@ -1,4 +1,8 @@
-import { eslintProvider, nodeProvider } from '@halfdomelabs/core-generators';
+import {
+  eslintProvider,
+  nodeProvider,
+  prettierProvider,
+} from '@halfdomelabs/core-generators';
 import {
   createProviderType,
   createGeneratorWithChildren,
@@ -6,6 +10,7 @@ import {
   copyFileAction,
 } from '@halfdomelabs/sync';
 import path from 'path';
+import * as prettierPluginTailwindcss from 'prettier-plugin-tailwindcss';
 import { z } from 'zod';
 
 import { reactProvider } from '../react/index.js';
@@ -28,17 +33,20 @@ const ReactTailwindGenerator = createGeneratorWithChildren({
     react: reactProvider,
     node: nodeProvider,
     eslint: eslintProvider,
+    prettier: prettierProvider,
   },
   exports: {
     reactTailwind: reactTailwindProvider,
   },
-  createGenerator({ globalBodyClasses }, { node, react, eslint }) {
+  createGenerator({ globalBodyClasses }, { node, react, eslint, prettier }) {
     const srcFolder = react.getSrcFolder();
+
+    const prettierPluginTailwindcssVersion = '0.5.14';
 
     node.addDevPackages({
       autoprefixer: '10.4.14',
       tailwindcss: '3.3.2',
-      'prettier-plugin-tailwindcss': '0.5.4',
+      'prettier-plugin-tailwindcss': prettierPluginTailwindcssVersion,
       '@tailwindcss/forms': '0.5.3',
     });
 
@@ -49,6 +57,12 @@ const ReactTailwindGenerator = createGeneratorWithChildren({
         'postcss.config.js',
         'tailwind.config.js',
       ]);
+
+    prettier.addPlugin({
+      name: 'prettier-plugin-tailwindcss',
+      version: prettierPluginTailwindcssVersion,
+      default: prettierPluginTailwindcss,
+    });
 
     react.getIndexFile().addCodeBlock('IMPORTS', "import './index.css'");