+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @halfdomelabs/tools
 
+## 0.5.4
+
+### Patch Changes
+
+- 922f0bf9: Upgrade prettier plugins and fix bug when prettier plugins were not present
+
+## 0.5.3
+
+### Patch Changes
+
+- f69fbf50: Upgrade vite to 5.2.4 and vitest to 1.4.0
+- 4c4cf8e5: Added sorting of package.json with prettier
+
 ## 0.5.2
 
 ### Patch Changes